backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1
slowapi==0.1.9

python-json-logger>=2.0.7
jinja2>=3.1.4
aiofiles>=23.2.1